Create download_equity_index_constituents.py
If your are coding a trading algorithm that invest on certain stocks that belong to an equity index and your willingness is outperform it, you must make sure that your backtesting uses the stock constituents that were part of the equity index on the starting date of your backtesting. If you use current index constituents, your algorithm backtesting results will be affected by survivorship bias. 
On this file, you can learn how to download S&P index constituents on a specific date.